sensation library websockets

+++ KeyboardControlledStimuli/Configurations.cpp
@@ -3,6 +3,6 @@
 namespace RandomWalk::Configurations {
     int Configurations::std_frequency = 256;
     float Configurations::std_intensity = 1.f;
-    int Configurations::std_duration = 256;
+    float Configurations::std_duration = 256.f;
     Vector3 Configurations::std_offset = Vector3({ -20.f, 0.f, 0.f });
 }
